chore(cli): drop now-unused stubRunE + clear two lint additions
After wiring exec/debug/watch/ps--watch to real RPCs, stubRunE has no
callers — delete it (and its `errors`/`os` imports). Also: fix the
unnecessary int32 conversion in server.Debug (inst.PID is already
int32) and acknowledge the discarded error from the ANSI clear-screen
write in runPsWatch.

Net effect on the golangci-lint baseline: 77 → 74 issues (the 3
I introduced or made obsolete this round). The remaining 70 are
pre-existing.

+++ cli/internal/daemon/server/server.go
@@ -1138,7 +1138,7 @@ func (s *Server) Debug(_ context.Context, req *connect.Request[anovelv1.DebugReq
 		return nil, connect.NewError(connect.CodeFailedPrecondition,
 			errors.New("debug: only go-exec targets can be attached to (container-mode dlv requires in-image dlv + port-forward)"))
 	}
-	port := int32(inst.PID + 20000)
+	port := inst.PID + 20000
 	hint := fmt.Sprintf(
 		"Attach with:\n  dlv attach %d --listen=:%d --headless --api-version=2\nThen connect from your editor (vscode: 'Connect to server' on port %d).",
 		inst.PID, port, port,
